services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["SakinaOuisrani <s.ouisrani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^3.0.0"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
